requirements.txt
# Add your requirements here like:
# numpy
# scipy>=0.9
pyyaml
boto
